Introduce deferrable checks and cases (#2)
This commit should close issue #2, asking for allowing for deferred evaluation of arguments in both checks and cases.

The syntax is as follows for checks:

    @case(4)
    @equals.defer(lambda: 2*4)
    def double(x):
        return 2*x

and similar for cases:

    @case.defer(lambda: 2*3*5)
    @equals([2,3,5])
    def get_prime_factors(x):
        ...

Deferred functions will pass non-lambda arguments through as-is, allowing for mixed deferred and non-deferred arguments.

The mechanism is allowed by a `@deferrable` decorator that adds a deferred version as a property of the passed function.
This is transparent to the user or developer: every check is deferrable by default, no extra action is needed.