New check model
This commit introduces a new model for creating checks.
The old way required to create a class for each new kind of check, and then calling the `decorator` class method on it.
The new way only required defining a function with a `checkify` decorator which returns the proper check kind.