Refactor check model (again)
Changes to the check and checklist model once more to suit them better for the coming modifier system.
The idea is to channel the behavior of checks and checklists through *attributes* instead of plain methods.
This allows taking this attributes and decorating/modifying them.

The most important change is that the Check class now is "streamlined" through the syntactic hell that is the decorating pattern I've chosen for Pycky.